init commit

+++ HospitalManagementSystem/database/PersonDAO.java
@@ -58,126 +58,6 @@ public class PersonDAO {
         }
         return false;
     }
-
-    public ArrayList<Person> getAllPeople() {
-        ArrayList<Person> list = new ArrayList<>();
-        String sql = "SELECT * FROM people";
-
-        Connection con = DatabaseConnection.getConnection();
-        if (con == null) return list;
-
-        try {
-            PreparedStatement statement = con.prepareStatement(sql);
-            ResultSet resultSet = statement.executeQuery();
-
-            while (resultSet.next()) {
-                list.add(extractPerson(resultSet));
-            }
-
-            resultSet.close();
-            statement.close();
-        }
-        catch (SQLException e) {
-            e.printStackTrace();
-        }
-        finally {
-            DatabaseConnection.closeConnection(con);
-        }
-        return list;
-    }
-    public ArrayList<Patient> getAllPatients() {
-        ArrayList<Patient> list = new ArrayList<>();
-        String sql = "SELECT * FROM people WHERE type='patient'";
-
-        Connection con = DatabaseConnection.getConnection();
-        if (con == null) return list;
-
-        try {
-            PreparedStatement statement = con.prepareStatement(sql);
-            ResultSet resultSet = statement.executeQuery();
-
-            while (resultSet.next()) {
-                list.add(new Patient(
-                        resultSet.getInt("id"),
-                        resultSet.getString("name"),
-                        resultSet.getInt("age"),
-                        resultSet.getString("department"),
-                        resultSet.getString("illness")
-                ));
-            }
-
-            resultSet.close();
-            statement.close();
-        }
-        catch (SQLException e) {
-            e.printStackTrace();
-        }
-        finally {
-            DatabaseConnection.closeConnection(con);
-        }
-
-        return list;
-    }
-    public ArrayList<Doctor> getAllDoctors() {
-        ArrayList<Doctor> list = new ArrayList<>();
-        String sql = "SELECT * FROM people WHERE type='doctor'";
-
-        Connection con = DatabaseConnection.getConnection();
-        if (con == null) return list;
-
-        try {
-            PreparedStatement statement = con.prepareStatement(sql);
-            ResultSet resultSet = statement.executeQuery();
-
-            while (resultSet.next()) {
-                list.add(new Doctor(
-                        resultSet.getInt("id"),
-                        resultSet.getString("name"),
-                        resultSet.getInt("age"),
-                        resultSet.getString("department"),
-                        resultSet.getString("specialization"),
-                        resultSet.getInt("experience"),
-                        resultSet.getDouble("salary")
-                ));
-            }
-
-            resultSet.close();
-            statement.close();
-        }
-        catch (SQLException e) {
-            e.printStackTrace();
-        }
-        finally {
-            DatabaseConnection.closeConnection(con);
-        }
-
-        return list;
-    }
-    private Person extractPerson(ResultSet resultSet) throws SQLException {
-        String type = resultSet.getString("type");
-
-        if (type.equalsIgnoreCase("doctor")) {
-            return new Doctor(
-                    resultSet.getInt("id"),
-                    resultSet.getString("name"),
-                    resultSet.getInt("age"),
-                    resultSet.getString("department"),
-                    resultSet.getString("specialization"),
-                    resultSet.getInt("experience"),
-                    resultSet.getDouble("salary")
-            );
-        }
-        else {
-            return new Patient(
-                    resultSet.getInt("id"),
-                    resultSet.getString("name"),
-                    resultSet.getInt("age"),
-                    resultSet.getString("department"),
-                    resultSet.getString("illness")
-            );
-        }
-    }
-
     public boolean updatePatient(Patient p) {
         String sql = "UPDATE people SET name=?, age=?, department=?, illness=? WHERE id=? AND type='patient'";
 
@@ -230,8 +110,10 @@ public class PersonDAO {
         }
         return false;
     }
-
     public boolean deletePerson(int id) {
+        if (getPersonById(id) == null) {
+            return false;
+        }
         String sql = "DELETE FROM people WHERE id = ?";
         Connection con = DatabaseConnection.getConnection();
         if (con == null) {
@@ -250,9 +132,120 @@ public class PersonDAO {
             DatabaseConnection.closeConnection(con);
         }
         return false;
-
-
     }
+
+    public ArrayList<Person> getAllPeople() {
+        ArrayList<Person> list = new ArrayList<>();
+        String sql = "SELECT * FROM people";
+
+        Connection con = DatabaseConnection.getConnection();
+        if (con == null) return list;
+
+        try {
+            PreparedStatement statement = con.prepareStatement(sql);
+            ResultSet resultSet = statement.executeQuery();
+
+            while (resultSet.next()) {
+                list.add(extractPerson(resultSet));
+            }
+
+            resultSet.close();
+            statement.close();
+        }
+        catch (SQLException e) {
+            e.printStackTrace();
+        }
+        finally {
+            DatabaseConnection.closeConnection(con);
+        }
+        return list;
+    }
+    public ArrayList<Patient> getAllPatients() {
+        ArrayList<Patient> list = new ArrayList<>();
+        String sql = "SELECT * FROM people WHERE type='patient'";
+
+        Connection con = DatabaseConnection.getConnection();
+        if (con == null) return list;
+
+        try {
+            PreparedStatement statement = con.prepareStatement(sql);
+            ResultSet resultSet = statement.executeQuery();
+
+            while (resultSet.next()) {
+                Person person = extractPerson(resultSet);
+                if (person instanceof Patient) {
+                    list.add((Patient) person);
+                }
+            }
+
+            resultSet.close();
+            statement.close();
+        }
+        catch (SQLException e) {
+            e.printStackTrace();
+        }
+        finally {
+            DatabaseConnection.closeConnection(con);
+        }
+
+        return list;
+    }
+    public ArrayList<Doctor> getAllDoctors() {
+        ArrayList<Doctor> list = new ArrayList<>();
+        String sql = "SELECT * FROM people WHERE type='doctor'";
+
+        Connection con = DatabaseConnection.getConnection();
+        if (con == null) return list;
+
+        try {
+            PreparedStatement statement = con.prepareStatement(sql);
+            ResultSet resultSet = statement.executeQuery();
+
+            while (resultSet.next()) {
+                Person person = extractPerson(resultSet);
+                if (person instanceof Doctor) {
+                    list.add((Doctor) person);
+                }
+            }
+
+            resultSet.close();
+            statement.close();
+        }
+        catch (SQLException e) {
+            e.printStackTrace();
+        }
+        finally {
+            DatabaseConnection.closeConnection(con);
+        }
+
+        return list;
+    }
+
+    private Person extractPerson(ResultSet resultSet) throws SQLException {
+        String type = resultSet.getString("type");
+
+        if (type.equalsIgnoreCase("doctor")) {
+            return new Doctor(
+                    resultSet.getInt("id"),
+                    resultSet.getString("name"),
+                    resultSet.getInt("age"),
+                    resultSet.getString("department"),
+                    resultSet.getString("specialization"),
+                    resultSet.getInt("experience"),
+                    resultSet.getDouble("salary")
+            );
+        }
+        else {
+            return new Patient(
+                    resultSet.getInt("id"),
+                    resultSet.getString("name"),
+                    resultSet.getInt("age"),
+                    resultSet.getString("department"),
+                    resultSet.getString("illness")
+            );
+        }
+    }
+
     public Person getPersonById(int id) {
         String sql = "SELECT * FROM people WHERE id=?";
 
@@ -261,8 +254,8 @@ public class PersonDAO {
 
         try {
             PreparedStatement statement = con.prepareStatement(sql);
-            ResultSet resultSet = statement.executeQuery();
             statement.setInt(1, id);
+            ResultSet resultSet = statement.executeQuery();
 
             if (resultSet.next()) return extractPerson(resultSet);
 
@@ -277,7 +270,6 @@ public class PersonDAO {
         }
         return null;
     }
-
     public ArrayList<Person> searchByName(String name) {
         ArrayList<Person> list = new ArrayList<>();
         String sql = "SELECT * FROM people WHERE name ILIKE ?";
@@ -288,7 +280,6 @@ public class PersonDAO {
         try {
             PreparedStatement statement = con.prepareStatement(sql);
             statement.setString(1, "%" + name + "%");
-
             ResultSet resultSet = statement.executeQuery();
 
             while (resultSet.next()) {
@@ -315,10 +306,8 @@ public class PersonDAO {
 
         try {
             PreparedStatement statement = con.prepareStatement(sql);
-
             statement.setDouble(1, min);
             statement.setDouble(2, max);
-
             ResultSet resultSet = statement.executeQuery();
 
             while (resultSet.next()) {
@@ -347,15 +336,10 @@ public class PersonDAO {
             ResultSet resultSet = statement.executeQuery();
 
             if (resultSet.next()) {
-                return new Doctor(
-                        resultSet.getInt("id"),
-                        resultSet.getString("name"),
-                        resultSet.getInt("age"),
-                        resultSet.getString("department"),
-                        resultSet.getString("specialization"),
-                        resultSet.getInt("experience"),
-                        resultSet.getDouble("salary")
-                );
+                Person person = extractPerson(resultSet);
+                if (person instanceof Doctor) {
+                    return (Doctor) person;
+                }
             }
 
             resultSet.close();
@@ -367,7 +351,6 @@ public class PersonDAO {
         finally {
             DatabaseConnection.closeConnection(con);
         }
-
         return null;
     }
 }
